Split up api functions and add session based data
Splitting up the functions makes the api easier to work with, and
switching to a model which uses session based data avoids one user
permanently changing the backend for another. This means the frontend
will need to generate some sort of unique_id which we probably want to
store in local storage.

Signed-off-by: Maxwell Henderson <mxwhenderson@gmail.com>